+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -53,6 +53,7 @@ use Phan\Language\Type\StringType;
 use Phan\Language\Type\VoidType;
 use Phan\Language\UnionType;
 
+use function end;
 use function implode;
 use function sprintf;
 
@@ -1444,7 +1445,7 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
 
     /**
      * @param Node $node
-     * A node to parse
+     * A node of kind ast\AST_CLOSURE or ast\AST_ARROW_FUNC to analyze
      *
      * @return Context
      * A new or an unchanged context resulting from
@@ -1463,6 +1464,15 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
             && !$return_type->hasType(NullType::instance(false))
         ) {
             $this->warnTypeMissingReturn($func, $node);
+        }
+        $uses = $node->children['uses'] ?? null;
+        // @phan-suppress-next-line PhanUndeclaredProperty
+        if (isset($uses->polyfill_has_trailing_comma) && Config::get_closest_minimum_target_php_version_id() < 80000) {
+            $this->emitIssue(
+                Issue::CompatibleTrailingCommaParameterList,
+                end($uses->children)->lineno ?? $uses->lineno,
+                ASTReverter::toShortString($node)
+            );
         }
         $this->analyzeNoOp($node, Issue::NoopClosure);
         $this->checkForFunctionInterfaceIssues($node, $func);
@@ -2973,7 +2983,16 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                 $parameters_seen[$parameter->getName()] = $i;
             }
         }
-        foreach ($node->children['params']->children as $param) {
+        $params_node = $node->children['params'];
+        // @phan-suppress-next-line PhanUndeclaredProperty
+        if (isset($params_node->polyfill_has_trailing_comma)) {
+            $this->emitIssue(
+                Issue::CompatibleTrailingCommaParameterList,
+                end($params_node->children)->lineno ?? $params_node->lineno,
+                ASTReverter::toShortString($node)
+            );
+        }
+        foreach ($params_node->children as $param) {
             $this->checkUnionTypeCompatibility($param->children['type']);
         }
         $this->checkUnionTypeCompatibility($node->children['returnType']);
@@ -3176,9 +3195,7 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
             }
             if ($argument_name_set) {
                 if ($argument === $argument_expression) {
-                    Issue::maybeEmit(
-                        $this->code_base,
-                        $this->context,
+                    $this->emitIssue(
                         Issue::PositionalArgumentAfterNamedArgument,
                         $argument->lineno ?? $node->lineno,
                         ASTReverter::toShortString($argument),
@@ -3194,11 +3211,17 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
         }
         // TODO: Make this a check that runs even without the $method object
         if ($has_unpack && $argument_name_set) {
-            Issue::maybeEmit(
-                $this->code_base,
-                $this->context,
+            $this->emitIssue(
                 Issue::ArgumentUnpackingUsedWithNamedArgument,
                 $node->lineno,
+                ASTReverter::toShortString($node)
+            );
+        }
+        // @phan-suppress-next-line PhanUndeclaredProperty
+        if (isset($node->polyfill_has_trailing_comma) && Config::get_closest_minimum_target_php_version_id() < 70300) {
+            $this->emitIssue(
+                Issue::CompatibleTrailingCommaArgumentList,
+                end($node->children)->lineno ?? $node->lineno,
                 ASTReverter::toShortString($node)
             );
         }