+++ src/Phan/Issue.php
@@ -564,6 +564,8 @@ class Issue
     public const CompatibleMatchExpression          = 'PhanCompatibleMatchExpression';
     public const CompatibleNullsafeOperator         = 'PhanCompatibleNullsafeOperator';
     public const CompatibleNamedArgument            = 'PhanCompatibleNamedArgument';
+    public const CompatibleTrailingCommaArgumentList = 'PhanCompatibleTrailingCommaArgumentList';
+    public const CompatibleTrailingCommaParameterList = 'PhanCompatibleTrailingCommaParameterList';
 
     // Issue::CATEGORY_GENERIC
     public const TemplateTypeConstant       = 'PhanTemplateTypeConstant';
@@ -4868,6 +4870,25 @@ class Issue
                 "Cannot use named arguments before php 8.0 in argument ({CODE})",
                 self::REMEDIATION_B,
                 3035
+            ),
+            // NOTE: The fact that the native php-ast does not track trailing commas is by design.
+            // It exposes the information that php's implementation stores internally,
+            // and that information is not available because php itself does not need it.
+            new Issue(
+                self::CompatibleTrailingCommaParameterList,
+                self::CATEGORY_COMPATIBLE,
+                self::SEVERITY_NORMAL,
+                "Cannot use trailing commas in parameter or closure use lists before php 8.0 in declaration of {FUNCTIONLIKE}. NOTE: THIS ISSUE CAN ONLY DETECTED BY THE POLYFILL.",
+                self::REMEDIATION_B,
+                3036
+            ),
+            new Issue(
+                self::CompatibleTrailingCommaArgumentList,
+                self::CATEGORY_COMPATIBLE,
+                self::SEVERITY_NORMAL,
+                "Cannot use trailing commas in argument lists before php 7.3 in {CODE}. NOTE: THIS ISSUE CAN ONLY DETECTED BY THE POLYFILL.",
+                self::REMEDIATION_B,
+                3037
             ),
 
             // Issue::CATEGORY_GENERIC