+++ tests/misc/fallback_test/.phan/config.php
@@ -14,6 +14,8 @@ use \Phan\Issue;
  * in addition to testing backwards compatibility checks and dead code detection.
  */
 return [
+    'minimum_target_php_version' => '7.2',
+
     // If true, missing properties will be created when
     // they are first seen. If false, we'll report an
     // error message.